move test into wrangler-remote-resources.test.ts
